start to update unit tests after splitting api
