+++ unit_tests/accounts_svc/main.py
@@ -1,10 +1,10 @@
 import unittest
 from test_application import TestApplication
 from test_threadsafe_configuration import TestThreadSafeConfiguration
-from test_sqlite_interface import TestSqliteInterface
+from test_da_user_data_access_layer import TestSqliteInterface
 from test_apis_authentication_api import TestAuthenticationAPI
 from test_apis_health_api import TestApiHealthApiView
-from test_extended_sql_interface import TestExtendedSqlInterface
+from test_services_authentication_service import TestExtendedSqlInterface
 
 if __name__ == "__main__":
     unittest.main()
